requirements.txt
appdirs==1.4.4
black==21.6b0
click==8.0.1
Django==2.0.7
django-cors-headers==2.5.0
django-filter==2.2.0
djangorestframework==3.10.0
importlib-metadata==4.5.0
mypy-extensions==0.4.3
pathspec==0.8.1
pytz==2021.1
regex==2021.4.4
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
zipp==3.4.1